Making sure everything is float32

+++ preprocess_save_patches_ISPRS.py
@@ -210,13 +210,13 @@ for i in tqdm(range(len(patches_tr))):
         # All multitasking labels are saved in one-hot
         # Segmentation
         np.save(os.path.join(folder_path, 'labels/seg', filename(i*5 + j)),
-                label_aug_h[j])
+                label_aug_h[j].astype(np.float32))
         # Boundary
-        bound_label_h = get_boundary_label(label_aug_h[j])
+        bound_label_h = get_boundary_label(label_aug_h[j]).astype(np.float32)
         np.save(os.path.join(folder_path, 'labels/bound', filename(i*5 + j)),
                 bound_label_h)
         # Distance
-        dist_label_h = get_distance_label(label_aug_h[j])
+        dist_label_h = get_distance_label(label_aug_h[j]).astype(np.float32)
         np.save(os.path.join(folder_path, 'labels/dist', filename(i*5 + j)),
                 dist_label_h)
         # Color
